Refactor text commands
