+++ modules/utils.py
@@ -1,0 +1,17 @@
+from LSP.plugin import LspTextCommand, Session
+
+from .constants import SESSION_NAME
+
+
+class LspJdtlsTextCommand(LspTextCommand):
+
+    session_name = SESSION_NAME
+
+    def run(self, edit, **args):
+        session = self.session_by_name(SESSION_NAME)
+        if not session:
+            return
+        self.run_jdtls_command(edit, session, **args)
+
+    def run_jdtls_command(self, edit, session: Session, **args):
+        ...
